requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.1
altair==5.5.0
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
apprise==1.9.5
asgi-lifespan==2.1.0
async-timeout==5.0.1
asyncpg==0.30.0
asyncssh==2.21.1
atpublic==6.0.2
attrs==25.4.0
billiard==4.2.2
blinker==1.9.0
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
cmdstanpy==1.3.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.2
coolname==2.2.0
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.71.0
dateparser==1.2.2
deprecation==2.1.0
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.24.8
dvc==3.63.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
dynaconf==3.2.12
entrypoints==0.4
evidently==0.7.15
exceptiongroup==1.3.0
Faker==37.12.0
fastapi==0.120.4
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.42.1
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.4
griffe==1.14.0
gto==1.9.0
h11==0.16.0
h2==4.3.0
holidays==0.83
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanize==4.14.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
influxdb-client==1.49.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.5.4
litestar==2.18.0
litestar-htmx==0.5.0
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mlflow==3.5.1
mlflow-skinny==3.5.1
mlflow-tracing==3.5.1
msgspec==0.19.0
multidict==6.7.0
multipart==1.3.0
mypy_extensions==1.1.0
narwhals==2.10.1
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
oauthlib==3.3.1
omegaconf==2.3.0
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
pathspec==0.12.1
patsy==1.0.2
pendulum==3.1.0
pillow==12.0.0
platformdirs==4.5.0
plotly==5.24.1
pluggy==1.6.0
polyfactory==2.22.3
prefect==3.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
prophet==1.2.1
protobuf==6.33.0
psutil==7.1.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
pywin32==311
PyYAML==6.0.3
reactivex==4.0.4
readchar==4.2.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rich-click==1.9.4
rpds-py==0.28.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scikit-learn==1.7.2
scipy==1.15.3
scmrepo==3.5.2
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
sqltrie==0.11.2
stanio==0.5.1
starlette==0.49.3
statsmodels==0.14.5
streamlit==1.51.0
tabulate==0.9.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.5.0
uuid6==2025.0.1
uv==0.9.7
uvicorn==0.38.0
vine==5.1.0
voluptuous==0.15.2
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
zc.lockfile==4.0
zipp==3.23.0
